adrv9371x: Use the output of IBUFDS_GTE2 as reference for the clock gens
In some cases (GTX2) the transceiver may gate the out_clk when it is in
reset. This will stop the clock generators from getting a clock prior
removing the reset of the XCVR.  The XCVR has a requirement of running
user clock while removing the reset. The correct sequence must be :

Enable device clocks (user clock)
Remove the reset from the XCVR